UI + gestion des permission + localisation + assets

+++ Weather/WeatherSys.swift
@@ -9,8 +9,8 @@ import Foundation
 
 class WeatherSys: Decodable {
     var country: String
-    var sunrise: Int
-    var sunset: Int
+    var sunrise: Double
+    var sunset: Double
 
     enum CodingKeys: String, CodingKey {
         case country
@@ -21,7 +21,7 @@ class WeatherSys: Decodable {
     required init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         country = try container.decode(String.self, forKey: .country)
-        sunrise = try container.decode(Int.self, forKey: .sunrise)
-        sunset = try container.decode(Int.self, forKey: .sunset)
+        sunrise = try container.decode(Double.self, forKey: .sunrise)
+        sunset = try container.decode(Double.self, forKey: .sunset)
     }
 }
